The beginning of Project
Additional Changes to do

1. Bankers

2. Sister Concern
Name of the company-
Registered address of the company-
Website of the company-
Logo (if website not available)-
If website and Logo of the sister concern is not available, then it may be excluded, but emphasis must be given in obtaining the website details.

3. Details of Departments (Clients)
The department name please add the website of the concerned department. Along with that if the company has other clients, please share those details as well.

4. Projects
Name of the Project-
Project Description- Place of Project, What is the project focused on solving, all other details available describing the project.
Start Date-(Month and Year)
Expected Completion Date-(Month and Year)
No. of people working in the project-(Approximate value if exact figure isn't available)

5. Images
The images of the Projects ongoing and have been completed are required.

6. Testimonials
An image of the employee, it can be a casual image, be it the best image of themselves they have taken, not necessarily a passport size photograph. Please keep in mind that the image would be displayed on company's website so it must be a good and presentable one.
His/Her designation.
His/Her feelings about the company.
It is also focused on explaining how his/her ambitions in life are fulfilled by working for the company and how they see at helping the company reach higher heights.
The Testimonies should not be of more than 70 words.
The best ones would be selected by the company and me and would be put on website.
I would initiate gathering the testimonial as a competition in the company's WhatsApp group.

7. Turnover of the company
The Turnover of the company is to be displayed on the website. So the turnover of the company as per the IT return is to be shared with me. It will be displayed as a percentage value.

8. Address and Contact Info